Add a DocumentStream abstraction
This abstraction represents a stream of documents to be processed, with
an implementation provided by the FileSorter sample application to
funnel files created in the watched inbox into the DocumentStream.

The sample implementation illustrates how to convert an event stream
into an IObservable, and how to convert an IEnumerable to an
IObservable.