Read all blob references from specified containers
Parse arguments into blob container references. For SAS connections, we
create a singleton array with the container reference for the given SAS.
For connection strings, we take the supplied container names and obtain
container references for each of them.

With the container references determined, we can then retrieve all the
blobs in those containers.