Add tests for FileSystemDocumentChannel
Test both initial contents of the directory without a watch, and
whilst watching the directory. The test files contain no data right now,
we just need to be able to enumerate files in a directory and react to
new files in that directory.

When testing for the initial contents of the directory, it is easier to
compare Enumerables of the directory contents than Observables,
especially as there is a fixed (and small) number of items in the test
directory to begin with.