Define the pipeline with a document source
The first thing that a document processing pipeline needs is the source
document for it to open and read the contents from.
`WithDocumentsFrom()` defines that part of the pipeline: the document
source.

This contains a naive implementation of this processing step which opens
the document and streams it to the Waives API for creation, wrapping up
the source document and the object created in the Waives API within a
WaivesDocument object. This object is the document which will be
processed within the remainder of the pipeline.