Add ObservableProcessExtension
This processes a sequence of WaivesDocuments, running an action on each. If the action succeeds, the document is included in the return observable. If the action throws an exception, the error action is run with that document wrapped in a ProcessingError and the document is not included in the return observable.