Add a PipelineException for fatal errors
Some processing errors are recoverable (e.g. if creating a document
fails, skip and move on to the next), but some are not (e.g.
authentication failed, so nothing will work). In the case of the
unrecoverable errors, we need to surface this to the calling application
by way of a PipelineException.