Add EventingDocumentSource and DocumentEmitter
This provides an easier API for event-based document sources such as the
FileSystemDocumentSource currently implemented by the file system
extension and used by the File Sorter sample app. It can be used to
implement either a pure event-based source, or a source requiring a
worker thread, such as running a query against a database on a regular
basis. It supports cancellation of long-running worker threads via the
standard CancellationToken object.