Implement a DocumentEmitter for filesystem watches
This wraps up most of the logic previously encapsulated in
`FileSystemDocumentSource.Create()` into the `DocumentEmitter`, whilst
also providing support for cancelling the watch. `EnableRaisingEvents`
proxies the equivalent property on `FileSystemWatcher` so that it can be
efficiently enabled and disabled if desired.

FileSystemDocumentSource is no longer a DocumentSource itself, and will
be renamed in time.