Use Document.ToString() in error messages
Do not use Document.SourceId in error messages, as this is an arbitrary
value from the client application which may not be suitable for use in
error messages (e.g. a database row ID won't provide much debugging
information). Instead use Document.ToString(), and encourage users to
provide their own implementation of this method.