Make logger configurable through CreatePipeline
If the logger is specified, use it, otherwise use a default ConsoleLogger which writes only Info messages and above so the user by default has some visibility into what is going on. Whichever logger is used, set the same logger for the underlying WaivesClient so we also get messages from that.