Replace cloing of requests with building from a template
Previously we cloned each HttpRequestMessage before sending, because we can't send each one more than once (if retrying). That was somewhat clunky and I was worried about the cloning code. So instead, pass the data needed to construct the HttpRequestMessage in an HttpRequestMessageTemplate and build the request as we're about to use it.