Pass ReliableRequestSender an ILogger
This has two effects, one good and one less good. The good thing is that
we can pass the logger from the factory method to the
ReliableRequestSender, allowing the ReliableRequestSender to encapsulate
its retry behaviour entirely. The bad thing is that we can't now pass an
arbitrary retry action to the ReliableRequestSender, but that feels like
an acceptable trade-off here.