Retry on failed requests, not thrown exceptions
The `HttpPolicyExtensions` configuration for Polly wasn't ever being
called because the responses it handles were being wrapped in
WaivesApiExceptions. The only part of the policy which was being used
was the handling of these exceptions. This commit reverses that
relationship, and removes handling of `WaivesApiExceptions` from the
policy, as these are exceptions that should be surfaced to the calling
application.