Add OnDocumentError handlers to sample apps
Neither the File Sorter nor the Blob Storage Processor sample apps
showed any example of using the OnDocumentError handler. This also meant
that the error messages thrown during processing of the pipeline were
never surfaced to the console and so it looked like the File Sorter was
sat doing nothing.