Change the type of ExtractionResult.Value to Object
The point of this property is that it is a strongly-typed representation of the field value (where a formatter is used in the extraction config), such as a DateTime or Decimal. Therefore as in the FPS SDK this should be an Object. I have tested that this works using Waives.Invoices.gb where the InvoiceDate Value property is set to a DateTime.