Move access token tests to new fixtures
The old Login() tests are now applicable to the AccessTokenService
class. New tests for the TokenFetchingRequestSender have been added to
verify the cache is correctly used. It turned out the cache was being
ignored because no Context was supplied, and the default cache key
strategy uses the Context's OperationKey value. This has been rectified
here too.