Give helpful error if document stream is empty
Previously, we sent the empty stream merrily on it's way to the Waives
API, where it was read and returned with a message stating "the body of
this request should be a binary file", which doesn't provide the
clearest indication of where the problem lies.

We now guard against null and empty document streams, and provide a more
appropriate message in these situations using `ArgumentException`s.