Create documents from non-empty unseekable streams
While the previous test bought us the ability to check for empty
unseekable streams, it left a gap in the test coverage for creating
documents from these streams. Here, if the supplied stream does not
support seeking, we wrap it in a FileBufferingReadStream (part of
ASP.NET Core, the WebUtilities reference added) so that we can seek on
that stream. This stream is configured to buffer any stream larger than
1MiB to disk, to a maximum of 30MB which is the largest size supported
by ASP.NET Core for requests.