Add LibLog to Waives.Http
LibLog's LogLevel enum is in conflict with the enum defined already as
part of the Waives.Http.Logging abstraction, so our own definition is
removed here.

The LIBLOG_PUBLIC compiler constant is defined so that the ILog
interface and other LibLog types are exposed publicly in the library as
an extension point for consumers.