Remove ref to ASP.NET WebApiClient
This is an unmaintained package now, and we were referencing only a
single method from it in the library (ReadAsAync<T>()). It depends on an
older version of Newtonsoft.Json which, in a .NET Framework project,
resolves to v6 or so. An implementation of ReadAsAsync<T>() is now
included in Waives.Http, and the latest version of Newtonsoft.Json is
referenced directly from the library.

Fixes #38.