Add support for redacting documents via Pipeline
As with the HTTP library, we redact based on results provided by the
specified extractor, and receive a Stream containing the redacted PDF.
In the Pipeline API, we need to keep the fluency of the pipeline
configuration, so `Pipeline.RedactWith()` additionally accepts an async
callback function to allow the calling application to perform some
operation with the resulting Stream (e.g. writing it to disk).